+++ tests/test_helpers/utils.py
@@ -16,6 +16,10 @@ def skip_if_env_var(var: str, exists=True):
     )
 
 
+def skip_if_no_groq(func):
+    return skip_if_env_var("GROQ_API_KEY", exists=False)(func)
+
+
 def skip_if_no_package(package):
     return pytest.mark.skipif(
         importlib.util.find_spec(package) is None,