+++ CHANGELOG.md
@@ -2,7 +2,8 @@
 
 ## v0.3.19 (Unreleased)
 
-- [vLLM model provider](https://inspect.ai-safety-institute.org.uk/models.html#sec-vllm) for higher performance use of Hugging Face and other local models.
+- [vLLM](https://inspect.ai-safety-institute.org.uk/models.html#sec-vllm) model provider.
+- [Groq](https://groq.com/) model provider.
 - Gracefully handle tool calls that include only a single value (rather than a named dict of parameters).
 - Support `tool_choice="any"` for OpenAI models (requires >= 1.24.0 of openai package).
 - Add `cwd` argument to `ToolEnvironment.exec()`.